requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
defusedxml==0.7.1
distro==1.9.0
dj-database-url==2.1.0
Django==5.2
django-cors-headers==4.7.0
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
factory-boy==3.3.0
Faker==24.4.0
docstring_parser==0.16
docx==0.2.4
drf-spectacular==0.28.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
instructor==1.7.2
Jinja2==3.1.5
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
oauthlib==3.2.2
openai==1.61.0
phonenumbers==9.0.5
pillow==11.1.0
PraisonAI==2.0.63
praisonaiagents==0.0.61
propcache==0.2.1
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.9.0
pyparsing==3.2.1
pytest==8.0.0
pytest-cov==4.1.0
pytest-django==4.8.0
pytest-factoryboy==2.6.0
pytest-mock==3.12.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.24.0
shellingham==1.5.4
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
tenacity==9.0.0
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.13.0
uritemplate==4.1.1
urllib3==2.3.0
yarl==1.18.3
